requirements.txt
flask==3.0.0
flask-cors==4.0.0
instagrapi>=2.1.0
google-generativeai==0.8.3
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
cryptography==41.0.3